refactor: move orchestration into monitor_app.main(), shim the entrypoint (#45)
