+++ monitor_app.py
@@ -65,3 +65,119 @@ def discovery_complete(found, window=DISCOVERY_WINDOW):
     if len(found) <= window:
         return False
     return found[-1] == found[-1 - window]
+
+
+def run_discovery(device_manager, max_wait=DISCOVERY_MAX_WAIT, window=DISCOVERY_WINDOW, sleep=time.sleep):
+    device_manager.update_devices()
+    logging.info("Starting discovery...")
+    device_manager.start_discovery()
+    found = []
+    waited = 0
+    while True:
+        sleep(1)
+        waited += 1
+        f = len(device_manager.devices())
+        logging.debug("Found {} BLE-devices so far".format(f))
+        found.append(f)
+        if discovery_complete(found, window) or waited >= max_wait:
+            break
+    device_manager.stop_discovery()
+    logging.info("Found {} BLE-devices".format(len(device_manager.devices())))
+
+
+def threaded_poller(dev, device_manager, logger_name, config, datalogger, queue):
+    logging.info("Trying to connect to {}...".format(dev.mac_address))
+    try:
+        device = SolarDevice(mac_address=dev.mac_address, manager=device_manager,
+                             logger_name=logger_name, config=config,
+                             datalogger=datalogger, queue=queue)
+    except Exception as e:
+        logging.error(e)
+        return
+    device.connect()
+
+
+def main(argv=None):
+    args = parse_args(argv)
+    ini_file = args.ini or "solar-monitor.ini"
+    try:
+        config = load_config(ini_file, adapter=args.adapter, debug=args.debug)
+    except ConfigError as e:
+        print(str(e))
+        return 1
+
+    debug = config.getboolean('monitor', 'debug', fallback=False)
+    if debug:
+        print("Debug enabled")
+    level = logging.DEBUG if debug else logging.INFO
+    duallog.setup('solar-monitor', minLevel=level, fileLevel=level, rotation='daily', keep=30)
+
+    pipeline = queue.Queue(maxsize=10000)
+    stop_event = threading.Event()
+    liveness = LivenessTracker()
+
+    try:
+        datalogger = DataLogger(config)
+    except Exception as e:
+        logging.error("Unable to set up datalogger")
+        logging.error(e)
+        return 1
+
+    executor = concurrent.futures.ThreadPoolExecutor(max_workers=100)
+    logger_future = executor.submit(run_logger, pipeline, datalogger, stop_event, liveness.record)
+
+    device_manager = SolarDeviceManager(adapter_name=config['monitor']['adapter'])
+    logging.info("Adapter status - Powered: {}".format(device_manager.is_adapter_powered))
+    if not device_manager.is_adapter_powered:
+        logging.info("Powering on the adapter ...")
+        device_manager.is_adapter_powered = True
+        logging.info("Powered on")
+
+    run_discovery(device_manager)
+
+    matched = 0
+    for dev in device_manager.devices():
+        logging.debug("Processing device {} {}".format(dev.mac_address, dev.alias()))
+        for section in config.sections():
+            if config.get(section, "mac", fallback=None) and config.get(section, "type", fallback=None):
+                mac = config.get(section, "mac").lower()
+                if dev.mac_address.lower() == mac:
+                    executor.submit(threaded_poller, dev, device_manager, section, config, datalogger, pipeline)
+                    liveness.expect(section)
+                    matched += 1
+                    logging.info("Waiting for device to connect")
+                    time.sleep(1)
+                    break
+
+    if matched == 0:
+        logging.error("No configured devices were discovered; exiting for restart.")
+        stop_event.set()
+        return 1
+
+    def _handle_signal(signum, _frame):
+        logging.info("Received signal %s; shutting down.", signum)
+        stop_event.set()
+        device_manager.stop()
+
+    signal.signal(signal.SIGTERM, _handle_signal)
+    signal.signal(signal.SIGINT, _handle_signal)
+
+    logging.debug("Waiting for devices to connect...")
+    time.sleep(10)
+
+    loop_thread = threading.Thread(target=device_manager.run, name="gatt-mainloop", daemon=True)
+    loop_thread.start()
+
+    logging.info("Supervising; terminate with Ctrl+C or SIGTERM")
+    exit_code = supervise(device_manager, logger_future, liveness, stop_event,
+                          check_interval=30.0, stale_timeout=600.0)
+
+    stop_event.set()
+    device_manager.stop()
+    for dev in device_manager.devices():
+        try:
+            dev.disconnect()
+        except Exception as e:
+            logging.warning("Error disconnecting %s: %s", getattr(dev, "mac_address", "?"), e)
+
+    return exit_code
